old_app/requirements.txt
psycopg2-binary==2.9.5
pandas==1.5.1
numpy==1.23.4